requirements.txt
psycopg2-binary==2.9.9
PyQt6==6.6.1
bcrypt==4.1.2
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0 